% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_exponential_probability}
\alias{calculate_exponential_probability}
\title{Calculate Probability Between Two Times for an Exponential Distribution}
\usage{
calculate_exponential_probability(lambda, time_start, time_end)
}
\arguments{
\item{lambda}{The rate parameter of the exponential distribution.}

\item{time_start}{The start time.}

\item{time_end}{The end time.}
}
\value{
The probability of the event occurring between time_start and time_end.
}
\description{
Calculates the probability of an event occurring between two times for an exponential distribution with a specified rate.
}
\examples{
calculate_exponential_probability(0.1, 2, 5)
}
