% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_probability_cost_less_than}
\alias{calculate_probability_cost_less_than}
\title{Calculate Probability of Cost Being Less Than a Given Value}
\usage{
calculate_probability_cost_less_than(mean_cost, sd_cost, threshold_cost)
}
\arguments{
\item{mean_cost}{The mean cost of the distribution.}

\item{sd_cost}{The standard deviation of the cost.}

\item{threshold_cost}{The cost threshold to compare against.}
}
\value{
The probability of the cost being less than the threshold.
}
\description{
Calculates the probability that a cost is less than a specified threshold, based on the mean and standard deviation of the cost.
}
\examples{
calculate_probability_cost_less_than(100, 15, 90)
}
