% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_probability_mean_less_than_or_equal}
\alias{calculate_probability_mean_less_than_or_equal}
\title{Calculate Probability of Sample Mean Less Than or Equal to Target}
\usage{
calculate_probability_mean_less_than_or_equal(
  population_mean,
  population_sd,
  sample_size,
  sample_mean
)
}
\arguments{
\item{population_mean}{The mean of the population.}

\item{population_sd}{The standard deviation of the population.}

\item{sample_size}{The size of the sample.}

\item{sample_mean}{The target sample mean.}
}
\value{
The probability as a numeric value.
}
\description{
Calculates the probability of the sample mean being less than or equal to a target value.
}
\examples{
calculate_probability_mean_less_than_or_equal(100, 15, 30, 95)
}
