% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_proportion_confidence_interval}
\alias{calculate_proportion_confidence_interval}
\title{Calculate Confidence Interval for a Proportion}
\usage{
calculate_proportion_confidence_interval(
  successes,
  sample_size,
  confidence_level
)
}
\arguments{
\item{successes}{The number of successes in the sample.}

\item{sample_size}{The size of the sample (an integer).}

\item{confidence_level}{The confidence level (e.g., 0.95 for 95\% confidence).}
}
\value{
A list with the lower and upper bounds of the confidence interval.
}
\description{
Calculates the confidence interval for a proportion based on the number of successes, sample size, and confidence level.
}
\examples{
calculate_proportion_confidence_interval(50, 100, 0.95)
}
