% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_sample_mean_probability}
\alias{calculate_sample_mean_probability}
\title{Calculate Sample Mean Probability}
\usage{
calculate_sample_mean_probability(
  population_mean,
  population_std_dev,
  sample_size,
  sample_mean_target
)
}
\arguments{
\item{population_mean}{The mean of the population.}

\item{population_std_dev}{The standard deviation of the population.}

\item{sample_size}{The size of the sample.}

\item{sample_mean_target}{The target sample mean.}
}
\value{
The probability as a numeric value.
}
\description{
Calculates the probability of obtaining a sample mean at least as extreme as a target value given a population mean, standard deviation, and sample size.
}
\examples{
calculate_sample_mean_probability(100, 15, 30, 105)
}
