% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_sample_size_for_proportion}
\alias{calculate_sample_size_for_proportion}
\title{Calculate Required Sample Size for Confidence Interval of a Proportion}
\usage{
calculate_sample_size_for_proportion(
  proportion,
  confidence_level,
  margin_of_error
)
}
\arguments{
\item{proportion}{The proportion of the outcome in the population (numeric, should be between 0 and 1).}

\item{confidence_level}{The desired confidence level (numeric, e.g., 0.99 for 99\% confidence).}

\item{margin_of_error}{The acceptable margin of error for the confidence interval (numeric, should be greater than 0).}
}
\value{
The calculated sample size as an integer (numeric).
}
\description{
This function calculates the minimum sample size needed to estimate the proportion
of a certain outcome in a population with a specified confidence level and margin of error.
This is essential for designing surveys and experiments to ensure that the estimates
fall within the desired accuracy.
}
\details{
The sample size is calculated using the following formula:
n = (z * sqrt(p * (1 - p)) / E)^2

Where:
\itemize{
\item z is the z-score corresponding to the desired confidence level,
\item p is the estimated proportion of the outcome,
\item E is the acceptable margin of error.
}
}
\examples{
calculate_sample_size_for_proportion(0.10, 0.99, 0.13)  # Example calculation

}
