% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cor3var}
\alias{cor3var}
\title{cor3var: Calculate the correlation coefficient from standard deviations and slope}
\usage{
cor3var(s_x, s_y, slope)
}
\arguments{
\item{s_x}{Standard deviation of x (numeric).}

\item{s_y}{Standard deviation of y (numeric).}

\item{slope}{The slope of the regression line (numeric).}
}
\value{
The correlation coefficient (r) (numeric).
}
\description{
This function calculates the correlation coefficient (r) using the standard
deviation of x (\( s_x \)), the standard deviation of y (\( s_y \)), and the slope of the
regression line. The correlation coefficient indicates the strength and direction
of the linear relationship between two variables.
}
\details{
The calculation is based on the following formula:
r = (slope * s_x) / s_y
}
\examples{
cor3var(1.5, 2.5, 0.8)

}
