% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{expected_value}
\alias{expected_value}
\title{expected_value: Calculate the expected value given win/lose amounts and probabilities}
\usage{
expected_value(win_amount, lose_amount, prob_win, prob_loss)
}
\arguments{
\item{win_amount}{The amount won if the event is successful.}

\item{lose_amount}{The amount lost if the event is unsuccessful.}

\item{prob_win}{Probability of winning the event.}

\item{prob_loss}{Probability of losing the event.}
}
\value{
The expected value of the event.
}
\description{
This function calculates the expected value of an event given the win amount, lose amount,
probability of win, and probability of loss.
}
\examples{
expected_value(100, -50, 0.6, 0.4)

}
