% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find_x_given_probability}
\alias{find_x_given_probability}
\title{Find x such that P(X < x) = 0.97725}
\usage{
find_x_given_probability(mu, sigma, probability)
}
\arguments{
\item{mu}{The mean of the distribution.}

\item{sigma}{The standard deviation of the distribution.}

\item{probability}{The cumulative probability for which to find x (e.g., 0.97725).}
}
\value{
The value of x corresponding to the given cumulative probability.
}
\description{
This function calculates the value of x such that the cumulative probability P(X < x) equals a given probability.
}
\examples{
find_x_given_probability(mu = 82, sigma = 4, probability = 0.97725)

}
