% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find_x_value}
\alias{find_x_value}
\title{Find X Value for a Given Probability}
\usage{
find_x_value(mean, sd, probability)
}
\arguments{
\item{mean}{The mean of the distribution.}

\item{sd}{The standard deviation of the distribution.}

\item{probability}{The target probability.}
}
\value{
The \code{x} value corresponding to the specified probability.
}
\description{
Finds the value of \code{x} such that the probability of a normally distributed random variable being greater than \code{x} is equal to the specified probability.
}
\examples{
find_x_value(100, 15, 0.05)
}
