% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gamma_with_mean_and_var}
\alias{gamma_with_mean_and_var}
\title{Compute Parameters and Probability for Gamma Distribution with Mean and Variance}
\usage{
gamma_with_mean_and_var(mu, sigma2, threshold = 12)
}
\arguments{
\item{mu}{The mean of the distribution (in millions of kilowatt-hours).}

\item{sigma2}{The variance of the distribution (in millions of kilowatt-hours squared).}

\item{threshold}{The threshold value for which the probability of exceeding is calculated (default is 12).}
}
\value{
A list containing:
\itemize{
\item \code{alpha}: The shape parameter of the gamma distribution.
\item \code{beta}: The scale parameter of the gamma distribution.
\item \code{prob_exceed_threshold}: The probability that the power consumption exceeds the threshold.
}
}
\description{
This function calculates the parameters (alpha and beta) of the gamma distribution
based on the provided mean and variance. It also computes the probability that the
daily power consumption exceeds a specified threshold (12 million kilowatt-hours).
}
\details{
The gamma distribution has the following characteristics:
\itemize{
\item The mean is equal to the product of alpha and beta.
\item The variance is equal to alpha times the square of beta.
}

The function solves for the values of alpha (shape parameter) and beta (scale parameter),
and then calculates the probability that the power consumption exceeds the threshold using the cumulative
distribution function (CDF) of the gamma distribution.
}
\examples{
result <- gamma_with_mean_and_var(mu = 6, sigma2 = 12)
print(result)

}
