% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{list_functions}
\alias{list_functions}
\title{list_functions: List all available functions in the package}
\usage{
list_functions()
}
\value{
A character vector of function names.
}
\description{
This function lists all the available functions in the 'statHelper' package.
}
\examples{
list_functions()

}
