% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mean_and_sd_Z}
\alias{mean_and_sd_Z}
\title{Mean and Standard Deviation for Z}
\usage{
mean_and_sd_Z()
}
\value{
A list containing the mean and standard deviation of Z.
}
\description{
This function returns the mean and standard deviation for the standard normal random variable Z,
which is always 0 for the mean and 1 for the standard deviation.
}
\examples{
mean_and_sd_Z()

}
