% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_normal_and_empirical}
\alias{plot_normal_and_empirical}
\title{This function plots the normal distribution for a given mean and standard deviation.
It also answers the question: According to the Empirical Rule, what are the bounds for
the middle 68\% of the data?}
\usage{
plot_normal_and_empirical(
  mu,
  sigma,
  x_range = c(mu - 4 * sigma, mu + 4 * sigma)
)
}
\arguments{
\item{mu}{The mean of the distribution.}

\item{sigma}{The standard deviation of the distribution.}

\item{x_range}{The range of x-axis values for plotting.}
}
\value{
A plot of the normal distribution and the middle 68\% range.
}
\description{
This function plots the normal distribution for a given mean and standard deviation.
It also answers the question: According to the Empirical Rule, what are the bounds for
the middle 68\% of the data?
}
\examples{
plot_normal_and_empirical(mu = 82, sigma = 4)

}
