% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prob_between_73_and_84}
\alias{prob_between_73_and_84}
\title{Find Probability P(73 < X < 84)}
\usage{
prob_between_73_and_84(mu, sigma, lower, upper)
}
\arguments{
\item{mu}{The mean of the distribution.}

\item{sigma}{The standard deviation of the distribution.}

\item{lower}{The lower bound of the interval.}

\item{upper}{The upper bound of the interval.}
}
\value{
The probability P(lower < X < upper).
}
\description{
This function calculates the probability that a normal random variable X lies between two values.
}
\examples{
prob_between_73_and_84(mu = 82, sigma = 4, lower = 73, upper = 84)

}
