% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prob_exact_bounds}
\alias{prob_exact_bounds}
\title{Find Probability P(X = upper and lower bounds)}
\usage{
prob_exact_bounds(mu, sigma, lower, upper)
}
\arguments{
\item{mu}{The mean of the distribution.}

\item{sigma}{The standard deviation of the distribution.}

\item{lower}{The lower bound.}

\item{upper}{The upper bound.}
}
\value{
The probability P(X = lower and upper).
}
\description{
This function calculates the probability that a normal random variable X is exactly at the lower and upper bounds.
}
\examples{
prob_exact_bounds(mu = 82, sigma = 4, lower = 73, upper = 84)

}
