% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prob_greater_than_79}
\alias{prob_greater_than_79}
\title{Find Probability P(X > 79)}
\usage{
prob_greater_than_79(mu, sigma, x)
}
\arguments{
\item{mu}{The mean of the distribution.}

\item{sigma}{The standard deviation of the distribution.}

\item{x}{The value to calculate the probability for.}
}
\value{
The probability P(X > x).
}
\description{
This function calculates the probability that a normal random variable X is greater than a given value.
}
\examples{
prob_greater_than_79(mu = 82, sigma = 4, x = 79)

}
