% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prob_greater_than_or_equal_x}
\alias{prob_greater_than_or_equal_x}
\title{Find Probability P(X >equals x)}
\usage{
prob_greater_than_or_equal_x(mu, sigma, x)
}
\arguments{
\item{mu}{The mean of the distribution.}

\item{sigma}{The standard deviation of the distribution.}

\item{x}{The value to calculate the probability for.}
}
\value{
The probability P(X >equals x).
}
\description{
This function calculates the probability that a normal random variable X is greater than or equal to a given value.
}
\examples{
prob_greater_than_or_equal_x(mu = 82, sigma = 4, x = 79)

}
