% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prob_less_than_83}
\alias{prob_less_than_83}
\title{Find Probability P(X < 83)}
\usage{
prob_less_than_83(mu, sigma, x)
}
\arguments{
\item{mu}{The mean of the distribution.}

\item{sigma}{The standard deviation of the distribution.}

\item{x}{The value to calculate the probability for.}
}
\value{
The probability P(X < x).
}
\description{
This function calculates the probability that a normal random variable X is less than a given value.
}
\examples{
prob_less_than_83(mu = 82, sigma = 4, x = 83)

}
