% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prob_Xbar_between_bounds}
\alias{prob_Xbar_between_bounds}
\title{Find P(lower <equals Xbar <equals upper)}
\usage{
prob_Xbar_between_bounds(
  lower,
  upper,
  population_mean,
  population_variance,
  sample_size
)
}
\arguments{
\item{lower}{The lower bound for (Xbar) (numeric).}

\item{upper}{The upper bound for (Xbar) (numeric).}

\item{population_mean}{The population mean (numeric).}

\item{population_variance}{The population variance (numeric).}

\item{sample_size}{The sample size (numeric).}
}
\value{
The probability P(lower <equals (Xbar) <equals upper).
}
\description{
This function calculates the probability that the sample mean (Xbar) lies between two bounds, lower and upper,
assuming the sample mean follows a normal distribution.
}
\examples{
prob_Xbar_between_bounds(45, 74, 67, 36, 100)

}
