% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prob_Xbar_less_than_bound}
\alias{prob_Xbar_less_than_bound}
\title{Find P(Xbar < bound)}
\usage{
prob_Xbar_less_than_bound(
  bound,
  population_mean,
  population_variance,
  sample_size
)
}
\arguments{
\item{bound}{The value of the bound for Xbar (numeric).}

\item{population_mean}{The population mean (numeric).}

\item{population_variance}{The population variance (numeric).}

\item{sample_size}{The sample size (numeric).}
}
\value{
The probability P(Xbar) < bound).
}
\description{
This function calculates the probability that the sample mean Xbar is less than a specified bound,
assuming the sample mean follows a normal distribution.
}
\examples{
prob_Xbar_less_than_bound(70, 67, 36, 100)

}
