% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prob_Z_less_than_target}
\alias{prob_Z_less_than_target}
\title{Find Probability P(Z < target)}
\usage{
prob_Z_less_than_target(target)
}
\arguments{
\item{target}{The value of Z for which to calculate the probability.}
}
\value{
The probability P(Z < target).
}
\description{
This function calculates the probability that the standard normal variable Z is less than a given target.
}
\examples{
prob_Z_less_than_target(1.2)

}
