% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{r_squared}
\alias{r_squared}
\title{r_squared: Calculate the coefficient of determination (R^2) given the correlation coefficient (r)}
\usage{
r_squared(r)
}
\arguments{
\item{r}{Correlation coefficient.}
}
\value{
The coefficient of determination (R^2).
}
\description{
This function calculates the coefficient of determination (R^2) from the correlation coefficient (r).
}
\examples{
r_squared(0.8)

}
