% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sample_proportion_prob}
\alias{sample_proportion_prob}
\title{Calculate Probability for Sample Proportions}
\usage{
sample_proportion_prob(
  population_proportion,
  sample_size,
  sample_proportion_target
)
}
\arguments{
\item{population_proportion}{The proportion in the population (e.g., 0.5).}

\item{sample_size}{The size of the sample (an integer).}

\item{sample_proportion_target}{The target sample proportion.}
}
\value{
The probability as a numeric value.
}
\description{
Calculates the probability for sample proportions based on a specified population proportion, sample size, and target sample proportion.
}
\examples{
sample_proportion_prob(0.5, 100, 0.55)
}
