% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sampling_distribution_mean_and_se}
\alias{sampling_distribution_mean_and_se}
\title{Calculate Mean and Standard Error of the Sampling Distribution}
\usage{
sampling_distribution_mean_and_se(
  population_mean,
  population_variance,
  sample_size
)
}
\arguments{
\item{population_mean}{The population mean (numeric).}

\item{population_variance}{The population variance (numeric).}

\item{sample_size}{The size of the sample (numeric).}
}
\value{
A list containing:
\itemize{
\item \code{sampling_mean}: The mean of the sampling distribution.
\item \code{standard_error}: The standard error of the sampling distribution.
}
}
\description{
This function calculates the mean and standard error of the sampling distribution of the sample mean.
Given the population mean and population variance, it computes the mean and standard error for a sample of size n.
}
\examples{
sampling_distribution_mean_and_se(67, 36, 100)

}
