% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{two_sample_t_test}
\alias{two_sample_t_test}
\title{Perform a Two-Sample T-Test for the Difference in Means}
\usage{
two_sample_t_test(mean_1, sd_1, n_1, mean_2, sd_2, n_2, alpha = 0.05)
}
\arguments{
\item{mean_1}{The mean of the first sample (numeric).}

\item{sd_1}{The standard deviation of the first sample (numeric).}

\item{n_1}{The sample size of the first sample (integer).}

\item{mean_2}{The mean of the second sample (numeric).}

\item{sd_2}{The standard deviation of the second sample (numeric).}

\item{n_2}{The sample size of the second sample (integer).}

\item{alpha}{The significance level for the test (default is 0.05).}
}
\value{
A list containing:
\itemize{
\item \code{t_statistic}: The t-test statistic (numeric).
\item \code{p_value}: The p-value for the hypothesis test.
\item \code{reject_H0}: A string indicating whether to reject or fail to reject the null hypothesis.
}
}
\description{
This function performs a two-sample t-test to determine if the mean of one group is greater than the mean of another group.
The function calculates the test statistic, the p-value, and provides a conclusion based on the significance level.
}
\examples{
result <- two_sample_t_test(10.40, 4.83, 97, 9.26, 4.68, 148, 0.05)
print(result)

}
