% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{uniform_probability_between}
\alias{uniform_probability_between}
\title{Calculate Probability Between Two Values for a Uniform Distribution}
\usage{
uniform_probability_between(lower_bound, upper_bound, x1, x2)
}
\arguments{
\item{lower_bound}{The lower bound of the uniform distribution (default is 1).}

\item{upper_bound}{The upper bound of the uniform distribution (default is 8).}

\item{x1}{The first value in the range.}

\item{x2}{The second value in the range.}
}
\value{
The percentage of observations falling between \code{x1} and \code{x2}.
}
\description{
This function calculates the probability (percent) that an observation falls between two given values
within the range of a uniform distribution defined between a lower and upper bound.
}
\examples{
uniform_probability_between(1, 8, 2, 5)  # Probability between 2 and 5
}
