% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cor3var}
\alias{cor3var}
\title{cor3var: Calculate the correlation coefficient from standard deviations and slope}
\usage{
cor3var(s_x, s_y, slope)
}
\arguments{
\item{s_x}{Standard deviation of x.}

\item{s_y}{Standard deviation of y.}

\item{slope}{The slope of the regression line.}
}
\value{
The correlation coefficient (r).
}
\description{
This function calculates the correlation coefficient (r) using the standard
deviation of x (s_x), the standard deviation of y (s_y), and the slope of the
regression line.
}
\examples{
cor3var(1.5, 2.5, 0.8)

}
