% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{least_squares_regression}
\alias{least_squares_regression}
\title{least_squares_regression: Perform least squares regression given ybar, xbar, s_x, s_y, and R}
\usage{
least_squares_regression(ybar, xbar, s_x, s_y, R)
}
\arguments{
\item{ybar}{Mean of y values.}

\item{xbar}{Mean of x values.}

\item{s_x}{Standard deviation of x values.}

\item{s_y}{Standard deviation of y values.}

\item{R}{Correlation coefficient between x and y.}
}
\value{
A list containing the slope and intercept of the regression line.
}
\description{
This function calculates the least squares regression line using the mean of y (ybar),
the mean of x (xbar), the standard deviations of x (s_x), and y (s_y), and the correlation coefficient (R).
}
\details{
The regression line is of the form: y = intercept + slope * x
}
\examples{
least_squares_regression(5, 3, 1.2, 2.1, 0.8)

}
