% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lm_summary}
\alias{lm_summary}
\title{lm_summary: Perform linear regression and show the summary}
\usage{
lm_summary(x, y)
}
\arguments{
\item{x}{A numeric vector of independent variable values.}

\item{y}{A numeric vector of dependent variable values.}
}
\value{
The summary of the linear model.
}
\description{
This function takes two data vectors x and y and fits a linear model (lm) to them. It then returns the summary of the model.
}
\examples{
lm_summary(c(1, 2, 3, 4), c(2, 4, 6, 8))

}
