% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prob_between}
\alias{prob_between}
\title{Calculate Probability Between Two Bounds}
\usage{
prob_between(lower_bound, upper_bound, mean, sd)
}
\arguments{
\item{lower_bound}{The lower bound.}

\item{upper_bound}{The upper bound.}

\item{mean}{The mean of the distribution.}

\item{sd}{The standard deviation of the distribution.}
}
\value{
The probability of the variable being between lower_bound and upper_bound.
}
\description{
Calculates the probability of a normally distributed variable being between two specified bounds.
}
\examples{
prob_between(10, 20, 15, 5)
}
