% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sample_mean_variance}
\alias{sample_mean_variance}
\title{sample_mean_variance: Calculate the mean and variance of a sample distribution}
\usage{
sample_mean_variance(X, P_X)
}
\arguments{
\item{X}{A numeric vector of values.}

\item{P_X}{A numeric vector of probabilities associated with X.}
}
\value{
A list containing the mean and variance of the sample distribution.
}
\description{
This function takes two vectors, X (values) and P(X) (probabilities), and calculates
the mean and variance of the sample distribution.
}
\examples{
sample_mean_variance(c(1, 2, 3), c(0.2, 0.5, 0.3))

}
